+++ src/handlers/vehicle_command.py
@@ -91,7 +91,9 @@ class VehicleCommandHandler:
                 command,
             )
 
-    async def handle_mqtt_command(self, *, topic: str, payload: str) -> None:
+    async def handle_mqtt_command(
+        self, *, topic: str, payload: str, retained: bool = False
+    ) -> None:
         analyzed_topic = self.__get_command_topics(topic)
         handler = self.__command_handlers.get(analyzed_topic.command_no_vin)
         if not handler:
@@ -103,7 +105,10 @@ class VehicleCommandHandler:
             )
         else:
             await self.__execute_mqtt_command_handler(
-                handler=handler, payload=payload, analyzed_topic=analyzed_topic
+                handler=handler,
+                payload=payload,
+                analyzed_topic=analyzed_topic,
+                retained=retained,
             )
 
     async def __execute_mqtt_command_handler(
@@ -112,19 +117,20 @@ class VehicleCommandHandler:
         handler: CommandHandlerBase,
         payload: str,
         analyzed_topic: _MqttCommandTopic,
+        retained: bool,
     ) -> None:
         topic = analyzed_topic.command_no_vin
         topic_no_global = analyzed_topic.command_no_global
         result_topic = analyzed_topic.response_no_global
 
         try:
-            execution_result = await handler.handle(payload)
+            execution_result = await handler.handle(payload, retained=retained)
             self.publisher.publish_str(result_topic, "Success")
             if execution_result.force_refresh:
                 self.vehicle_state.set_refresh_mode(
                     RefreshMode.FORCE, f"after command execution on topic {topic}"
                 )
-            if execution_result.clear_command:
+            if execution_result.clear_command and not retained:
                 self.publisher.clear_topic(topic_no_global)
         except MqttGatewayException as e:
             self.__report_command_failure(
@@ -145,14 +151,14 @@ class VehicleCommandHandler:
                 )
                 return
             try:
-                execution_result = await handler.handle(payload)
+                execution_result = await handler.handle(payload, retained=retained)
                 self.publisher.publish_str(result_topic, "Success")
                 if execution_result.force_refresh:
                     self.vehicle_state.set_refresh_mode(
                         RefreshMode.FORCE,
                         f"after command execution on topic {topic}",
                     )
-                if execution_result.clear_command:
+                if execution_result.clear_command and not retained:
                     self.publisher.clear_topic(topic_no_global)
             except Exception as retry_err:
                 self.__report_command_failure(